# Documentacion/openapi/full_api.yaml
openapi: 3.0.3
info:
  title: Plataforma - API (Auth, Invoices, OCR, SUNAT)
  description: >-
    API completa y profesional para la plataforma de facturación, OCR y
    verificación SUNAT. Incluye contratos para Auth, Invoices (CRUD,
    upload-pdf, apply-parsed, verify-ruc), OCR ingestion/status/suggestions y
    servicio SUNAT.
  version: 1.0.0
servers:
  - url: https://api.example.com/api
    description: URL base (reemplazar en cada entorno)
  - url: http://localhost:5000/api
    description: Entorno de desarrollo (usar el servidor local apropiado)
...
security:
  - bearerAuth: []
paths:
...
  /invoices:
    get:
      tags: [invoices]
      summary: Listar facturas (paginado, filtros)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: q
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [sale, purchase]
      responses:
        '200':
          description: Lista paginada de facturas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedInvoices'
    post:
      tags: [invoices]
      summary: Crear una factura
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceCreate'
      responses:
        '201':
          description: Factura creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          $ref: '#/components/responses/BadRequest'
  /invoices/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          format: uuid
          type: string
    get:
      tags: [invoices]
      summary: Obtener factura por id
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Factura
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [invoices]
      summary: Actualizar factura
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceUpdate'
      responses:
        '200':
          description: Factura actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          $ref: '#/components/responses/BadRequest'
    delete:
      tags: [invoices]
      summary: Eliminar factura
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Eliminado
...
  /ocr/invoices:
    post:
      tags: [ocr]
      summary: Ingest OCR (alternativa directa al upload desde invoices)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '202':
          description: Job encolado para OCR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobCreated'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: refreshToken
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
  responses:
    BadRequest:
      description: Petición inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
...
    JobCreated:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
...
    PaginatedInvoices:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Invoice'
        meta:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
    InvoiceLine:
      type: object
      properties:
        id:
          type: string
          format: uuid
        line_no:
          type: integer
        product_id:
          type: string
        description:
          type: string
        quantity:
          type: number
        unit_price:
          type: number
        tax_rate:
          type: number
        line_total:
          type: number
    Invoice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [sale, purchase]
        document_type:
          type: string
        document_subtype:
          type: string
        is_boleta:
          type: boolean
        invoice_number:
          type: string
        date:
          type: string
          format: date
        party_id:
          type: string
        party_ruc:
          type: string
        ruc:
          type: string
        party_name:
          type: string
        currency:
          type: string
        status:
          type: string
        sub_total:
          type: number
        tax_total:
          type: number
        total:
          type: number
        lines:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceLine'
        metadata:
          type: object
    InvoiceCreate:
      type: object
      required: [type, date, lines]
      properties:
        type:
          type: string
          enum: [sale, purchase]
        document_type:
          type: string
        document_number:
          type: string
        is_boleta:
          type: boolean
        ruc:
          type: string
        party_name:
          type: string
        currency:
          type: string
          default: USD
        lines:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceLine'
    InvoiceUpdate:
      type: object
      properties:
        status:
          type: string
        metadata:
          type: object
...
security:
  - bearerAuth: []